Remove existing thread state for clone children
When parsing clone exit events, specifically for the child half of a
clone and when in a container, always remove any existing thread state
that may be attached to the event and remove that state from the thread
manager.

This matches the logic on the parent side, which is to let the child
create the thread.

This ensures that if there does happen to be an existing thread state
from an old thread that happened to share the same thread id, that it
will be removed.

This fixes https://github.com/draios/sysdig/issues/664.